refactor(develop): 페이지 단위 컴포넌트 생성
