--- instagram/src/components/Main.jsx
@@ -1,13 +1,0 @@
-import Navigation from "./Navigation";
-import MainBoard from "./MainBoard";
-
-function Main() {
-  return (
-    <>
-      <Navigation />
-      <MainBoard />
-    </>
-  );
-}
-
-export default Main;
