+++ instagram/src/pages/MainPage.jsx
@@ -1,0 +1,13 @@
+import Navigation from "../components/Navigation";
+import MainBoard from "../components/MainBoard";
+
+function MainPage() {
+  return (
+    <>
+      <Navigation />
+      <MainBoard />
+    </>
+  );
+}
+
+export default MainPage;
